fix: run check_suite
