+++ packages/merge-on-green/README.md
@@ -1,19 +1,19 @@
-mergeOnGreen
+# merge-on-green
 
 Instructions are provided in [googleapis/repo-automation-bots](https://github.com/googleapis/repo-automation-bots/blob/master/README.md) for deploying and testing your bots.
 
 This bot uses nock for mocking requests to GitHub, and snap-shot-it for capturing responses; This allows updates to the API surface to be treated as a visual diff, rather than tediously asserting against each field.
 
-Running tests:
+## Running tests:
 
 `npm run test`
 
-To update snapshots:
+## To update snapshots:
 
 `npm run test:snap`
 
-Contributing
-If you have suggestions for how mergeOnGreen could be improved, or want to report a bug, open an issue! We'd love all and any contributions.
+## Contributing
+If you have suggestions for how merge-on-green could be improved, or want to report a bug, open an issue! We'd love all and any contributions.
 
 For more, check out the Contributing Guide.
